military_rag_qa_updated/requirements.txt
faiss-cpu
sentence-transformers
transformers
torch
